use custom react hook instead
- rename dateFormatString to formatDateString
